+++ CHANGELOG.md
@@ -22,6 +22,22 @@ All notable changes to this project will be documented here. For more details, v
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.4.0] - UNRELEASED
+
+### Added
+
+- `--tls.min-version` option
+- `--ca-certs` option
+- `--verify` option
+
+### Changed
+
+- Always use the CA certificates from the system trust store + CA bundle from Mozilla
+
+### Deprecated
+
+-- `--ssl-context` is deprecated and will be removed in a future version.
+
 ## [10.3.0] - 2026-07-28
 
 ### Added